Fix Problems
